+++ plone/app/content/browser/contents/rearrange.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from OFS.interfaces import IOrderedContainer
 from plone.app.content.browser.contents import ContentsBaseAction
 from plone.app.content.utils import json_loads
 from plone.folder.interfaces import IExplicitOrdering
@@ -15,6 +16,9 @@ class OrderContentsBaseAction(ContentsBaseAction):
         try:
             ordering = self.context.aq_base.getOrdering()
         except AttributeError:
+            if IOrderedContainer.providedBy(self.context):
+                # Archetype
+                return IOrderedContainer(self.context)
             return None
         if not IExplicitOrdering.providedBy(ordering):
             return None